week4/default/update: 포인트 타입 변경 (BigDecimal -> int)

+++ src/main/java/hanghaeplus/application/point/dto/PointResponse.java
@@ -1,11 +1,9 @@
 package hanghaeplus.application.point.dto;
 
-import java.math.BigDecimal;
-
 public class PointResponse {
 
     public record PointSelection (
-            BigDecimal point
+            int point
     ) {
 
     }
